requirements.txt
flask==3.1.0
flask-session==0.8.0
openai>=1.0.0
google-generativeai>=0.8.3
reportlab==4.2.5
yookassa>=3.3.0
python-dotenv==1.2.2
Pillow>=10.0.0
gunicorn==21.2.0
